refactor: migrate card list page
