+++ src/features/card/index.ts
@@ -1,7 +1,7 @@
-export { BackText } from "./components/BackText";
+export { BackText, type BackTextProps } from "./components/BackText";
+export { Card, type CardProps } from "./components/Card";
 export { CardOverlay } from "./components/CardOverlay";
 export { FrontText } from "./components/FrontText";
 export { CardFormContainer } from "./containers/CardFormContainer";
-export { CardListContainer } from "./containers/CardListContainer";
 export { CardViewContainer } from "./containers/CardViewContainer";
 export { useCardMutations } from "./hooks/useCardMutations";
